pyproject.toml
[project]
name = "ollamaandlanggraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "imap-client>=0.1.8",
    "langchain>=0.3.25",
    "langchain-ollama>=0.3.3",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
